chore: make reset_gitlab() better
Saw issues in the CI where reset_gitlab() would fail. It would fail to
delete the group that is created when GitLab starts up. Extending the
timeout didn't fix the issue.

Changed the code so now it will check if the item is deleted and if it
isn't it will call the delete() method again to ensure that GitLab
knows it should be deleted. Since making this change I have not been
able to reproduce the failure in reset_gitlab().

Also added some logging functionality that can be seen if logging is
turned on in pytest.